+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.36.0] - 2023-09-15
+### Changed
+- Responses coming from App Store Connect are now processed better
+- Error responses coming from App Store Connect are now processed better
+
 ## [0.35.0] - 2023-06-05
 ### Added
 - Initial release.